+++ client/src/components/CheckoutForm.js
@@ -1,4 +1,5 @@
 import React, { useState } from "react";
+import {useForm} from '../hooks/useForm'
 
 const initialValue = {
   firstName: "",
@@ -13,18 +14,29 @@ const initialValue = {
 // Build out the logic needed for a form custom hook (see the useForm.js file)
 // and replace the necessary stateful logic from CheckoutForm with the hook
 
-const CheckoutForm = (props) => {
-  const [showSuccessMessage, setShowSuccessMessage] = useState(false);
-  const [values, setValues] = useState(initialValue);
 
-  const handleChanges = (e) => {
-    setValues({ ...values, [e.target.name]: e.target.value });
-  };
+ function CheckoutForm() {
 
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    setShowSuccessMessage(true);
-  };
+  // const [showSuccessMessage, setShowSuccessMessage] = useState(false);
+
+  const [values,handleChanges,handleSubmit,] = useForm(initialValue);
+
+  
+// const CheckoutForm = (props) => {
+//   const [showSuccessMessage, setShowSuccessMessage] = useState(false);
+//   const [values, setValues] = useState(initialValue);
+
+  // const handleChanges = (e) => {
+  //   setValues({ ...values, [e.target.name]: e.target.value });
+  // };
+
+  // const handleSubmit = (e) => {
+  //   e.preventDefault();
+  //   setShowSuccessMessage(true);
+  // };
+
+
+
 
   return (
     <>
@@ -69,7 +81,7 @@ const CheckoutForm = (props) => {
         <button>Checkout</button>
       </form>
 
-      {showSuccessMessage && (
+      {useForm.showSuccessMessage && (
         <div className="success-message" data-testid="successMessage">
           <p>
             You have ordered some plants! Woo-hoo! <span role="img">🎉</span>
