+++ client/src/hooks/useLocalStorage.js
@@ -1,0 +1,21 @@
+import { useState } from "react";
+
+export const useLocalStorage = (key, initialValue) => {
+
+
+
+  const [storedValue, setStoredValue] = useState(() => {
+    const item = window.localStorage.getItem(key);
+    return item ? JSON.parse(item) : initialValue;
+  });
+
+  const setValue = value => {
+    // Save state
+    setStoredValue(value);
+    // Save to local storage
+    window.localStorage.setItem(key, JSON.stringify(value));
+  };
+
+  return [storedValue, setValue];
+  
+};
